add like unlike, search

+++ src/views/Pages/Register/Register.js
@@ -4,55 +4,72 @@ import axios from 'axios';
 import { connect } from 'react-redux';
 import { login } from '../../../actions/authAction';
 import URL from '../../../helpers/url';
-
+import NotificationAlert from 'react-notification-alert';
 const _url = 'http://localhost:10010';
 
 class Register extends Component {
 
   handleSignup = (e) => {
     e.preventDefault();
-    
     var username = this.username.value;
     var fullname = this.fullname.value;
     var password = this.password.value;
-    var self=this;
-
-    // this.props.form.validateFields((err, values) => {
-    //   if (!err) {
-        axios.post(_url + '/user/signup', {
-          user_info_signup: {
-            username: username,
-            fullname: fullname,
-            password: password
+    var confirmPassword = this.confirmPassword.value;
+    
+    var self = this;
+    if (password===confirmPassword) {
+      axios.post(_url + '/user/signup', {
+        user_info_signup: {
+          username: username,
+          fullname: fullname,
+          password: password
+        }
+      }).then(function (res) {
+        if (res.data.status === 200) {
+          const { username, token } = res.data;
+          var obj = { username, token };
+          self.props.login(obj);
+          localStorage.setItem("username", username);
+          localStorage.setItem("token", token);
+          self.props.history.push(URL.HOME);
+        }
+        else {
+          const options_noti = {
+            place: 'tr',
+            message: res.data.message,
+            type: 'danger',
+            autoDismiss: 3,
           }
-        }).then(function (res) {
-          if (res.data.status === 200) {
-            const { username, token } = res.data;
-            var obj = { username, token };
-            self.props.login(obj);
-            localStorage.setItem("username", username);
-            localStorage.setItem("token", token);
-            self.props.history.push(URL.HOME);
-          } 
+          self.refs.notificationAlert.notificationAlert(options_noti);
+        }
 
-        })
-          .catch(function (error) {
-            console.log(error);
-          });
-    //   }
-    // });
+      })
+        .catch(function (error) {
+          console.log(error);
+        });
+    }
+    else {
+      const options_pass = {
+        place: 'tr',
+        message: 'Two passwords that you enter is inconsistent!',
+        type: 'danger',
+        autoDismiss: 3,
+      }
+      self.refs.notificationAlert.notificationAlert(options_pass);
+    }
   }
 
 
   render() {
     return (
       <div className="app flex-row align-items-center">
+        <NotificationAlert ref="notificationAlert" />
         <Container>
           <Row className="justify-content-center">
             <Col md="6">
               <Card className="mx-4">
                 <CardBody className="p-4">
-                  <Form  onSubmit={this.handleSignup}>
+                  <Form onSubmit={this.handleSignup}>
                     <h1>Register</h1>
                     <p className="text-muted">Create your account</p>
                     <InputGroup className="mb-3">
@@ -61,7 +78,7 @@ class Register extends Component {
                           <i className="icon-user"></i>
                         </InputGroupText>
                       </InputGroupAddon>
-                      <Input innerRef={(node) => this.username = node} type="text" placeholder="Username" autoComplete="username" />
+                      <Input required innerRef={(node) => this.username = node} type="text" placeholder="Username" autoComplete="username" />
                     </InputGroup>
                     <InputGroup className="mb-3">
                       <InputGroupAddon addonType="prepend">
@@ -69,7 +86,7 @@ class Register extends Component {
                           <i className="fa fa-id-card-o "></i>
                         </InputGroupText>
                       </InputGroupAddon>
-                      <Input innerRef={(node) => this.fullname = node} type="text" placeholder="Fullname" autoComplete="fullname" />
+                      <Input required innerRef={(node) => this.fullname = node} type="text" placeholder="Fullname" autoComplete="fullname" />
                     </InputGroup>
                     <InputGroup className="mb-3">
                       <InputGroupAddon addonType="prepend">
@@ -77,7 +94,7 @@ class Register extends Component {
                           <i className="icon-lock"></i>
                         </InputGroupText>
                       </InputGroupAddon>
-                      <Input innerRef={(node) => this.password = node} type="password" placeholder="Password" autoComplete="new-password" />
+                      <Input required innerRef={(node) => this.password = node} type="password" placeholder="Password" autoComplete="new-password" />
                     </InputGroup>
                     <InputGroup className="mb-4">
                       <InputGroupAddon addonType="prepend">
@@ -85,7 +102,7 @@ class Register extends Component {
                           <i className="icon-lock"></i>
                         </InputGroupText>
                       </InputGroupAddon>
-                      <Input type="password" placeholder="Repeat password" autoComplete="new-password" />
+                      <Input required innerRef={(node) => this.confirmPassword = node} type="password" placeholder="Repeat password" autoComplete="new-password" />
                     </InputGroup>
                     <Button color="success" block>Create Account</Button>
                   </Form>
@@ -108,8 +125,8 @@ class Register extends Component {
     );
   }
 }
-const mapStateToProps = (state) => ({ 
-  auth: state.auth 
+const mapStateToProps = (state) => ({
+  auth: state.auth
 });
 
 const mapDispatchToProps = (dispatch) => ({
